feat: 🎸 delete header and add user and log out button in bottom
for feature add logic for button to log out and sign in admin user and
dispaying his name and icon. Sidebar have 4 pages for now, dashboard
when we create graps with data from database, products when we can make
new, update and delete already existing products. Orders and users same.

✅ Closes: SHRP-12